% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{getCompendium}
\alias{getCompendium}
\title{load all compendium data into a TreeSummarizedExperiment}
\usage{
getCompendium(bfc = BiocFileCache::BiocFileCache())
}
\arguments{
\item{bfc}{BiocFileCache object to use}
}
\value{
a `TreeSummarizedExperiment`
}
\description{
load all compendium data into a TreeSummarizedExperiment
}
\examples{
cpd <- get_compendium()

dim(cpd)
cpd
assayNames(cpd)
head(colData(cpd))

}
