% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{loadme}
\alias{loadme}
\title{load data}
\usage{
loadme(fname)
}
\arguments{
\item{fname}{The filename or URL of the `wide.csv` file.
The file can (and probably should) be gzipped.}
}
\description{
load data
}
