% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{get_compendium}
\alias{get_compendium}
\title{load data}
\usage{
get_compendium(bfc = BiocFileCache::BiocFileCache(), ...)
}
\arguments{
\item{bfc}{BiocFileCache object to use}

\item{...}{passe}
}
\description{
load data
}
