% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{get_compendium}
\alias{get_compendium}
\title{load all compendium data into a TreeSummarizedExperiment}
\usage{
get_compendium(bfc = BiocFileCache::BiocFileCache())
}
\arguments{
\item{bfc}{BiocFileCache object to use}
}
\description{
load all compendium data into a TreeSummarizedExperiment
}
\examples{

# The data are large, so avoid running in example

\dontrun{
cpd <- get_compendium()

dim(cpd)
cpd
assayNames(cpd)
head(colData(cpd))
}

}
